# general CCL optimizations for some Llama ops