--- a/models/demos/llama3_subdevices/tests/tg_perf_unit_tests/test_ccl_async_perf_TG_llama.py
+++ b/models/demos/llama3_subdevices/tests/tg_perf_unit_tests/test_ccl_async_perf_TG_llama.py
@@ -15,7 +15,7 @@ THRESHOLD = 0.7
     "ag_type, warmup_iters, perf_target_us",
     [
         ("sdpa", 15, 8.95),
-        ("binary_mult", 15, 12.9),
+        ("binary_mult", 15, 10.07),
         ("layernorm", 15, 6.26),
     ],
 )