--- a/tt_metal/fabric/hw/inc/edm_fabric/fabric_connection_manager.hpp
+++ b/tt_metal/fabric/hw/inc/edm_fabric/fabric_connection_manager.hpp
@@ -44,6 +44,7 @@ public:
         }
     }
 
+    template <bool connect = false, bool wait_for_connection_open_finish = false>
     static FabricConnectionManager build_from_args(std::size_t& arg_idx) {
         FabricConnectionManager connection_manager;
         connection_manager.connection_flags = static_cast<uint8_t>(get_arg_val<uint32_t>(arg_idx++) != 0)
@@ -51,14 +52,38 @@ public:
         if (connection_manager.has_forward_connection()) {
             connection_manager.forward_fabric_sender =
                 tt::tt_fabric::WorkerToFabricEdmSender::build_from_args<ProgrammableCoreType::TENSIX>(arg_idx);
+            if constexpr (connect) {
+                connection_manager.forward_fabric_sender.open_start();
+            }
         }
         connection_manager.connection_flags |= static_cast<uint8_t>(get_arg_val<uint32_t>(arg_idx++) != 0)
                                                << BACKWARD_CONNECTION_FLAG_OFFSET;
         if (connection_manager.has_backward_connection()) {
             connection_manager.backward_fabric_sender =
                 tt::tt_fabric::WorkerToFabricEdmSender::build_from_args<ProgrammableCoreType::TENSIX>(arg_idx);
+            if constexpr (connect) {
+                connection_manager.backward_fabric_sender.open_start();
+            }
+        }
+
+        if constexpr (connect && wait_for_connection_open_finish) {
+            if (connection_manager.has_forward_connection()) {
+                connection_manager.forward_fabric_sender.open_finish();
+            }
+            if (connection_manager.has_backward_connection()) {
+                connection_manager.backward_fabric_sender.open_finish();
+            }
         }
         return connection_manager;
+    }
+
+    inline void open_finish() {
+        if (has_forward_connection()) {
+            forward_fabric_sender.open_finish();
+        }
+        if (has_backward_connection()) {
+            backward_fabric_sender.open_finish();
+        }
     }
 
     tt::tt_fabric::WorkerToFabricEdmSender& get_forward_connection() {