### What's changed
Applied some minor optimizations in all-gather an all-reduce kernels and updated perf bounds && thresholds for ops.

Deltas over perf thresholds in main are quite substantial but those numbers on main are slightly stale (and conservative). The PR is showing substantial improvement (e.g. >= 4us in some cases), but in reality the improvement was not that substantial due to outdated numbers. In practice, the real improvement is in the 500-1500ns range.


The following table shows the _real_ improvements (you may notice the "time before" here is lower than the targets listed in the tests before this PR.
## (IGNORE -- INVALID) Previous perf before and after 
| Op Instance | Time Before (us) | Time After (us) | ~Savings (us) |
|-|-|-|-|
| All-reduce ff2 | 18.156 | 17.6 | 0.550 |
| All-reduce ff1 | 18.667 | 18.0 | 0.667 |
| All-reduce qkv | 11.65 | 11.5 | 0.15 |
| all-reduce lm_head | 58.4 | 58.4 | No Change|
| all-gather sdpa | 10.1 | 8.95 | 1.105 |
| all-gather binary_mult | 12.9 | 10.07 | 2.2 |
| all-gather layernorm | 7.03 | 6.26 | 0.78 |

## Corrected perf before and after
| Op Instance | Time Before (us) | Time After (us) | ~Savings (us) |
|-|-|-|-|
| all-reduce ff1     | 19.8  | 19.18 | 0.620 |
| all-reduce ff2     | 19.49 | 18.55 | 0.851 |
| all-reduce qkv     | 12.2  | 11.9 | 0.3 |
| all-reduce lm_head | 62.15 | 61.82| 0.33 |
| all-gather binarymult   | 11.85 | 10.54 | 0.81 |
| all-gather layernorm   | 7.40  | 6.47 | 0.93 |
| all-gather sdpa    | 10.78 | 9.49  | 1.29 |


### Checklist
- [x] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI: https://github.com/tenstorrent/tt-metal/actions/runs/14088834759
- [ ] TG Model Perf: https://github.com/tenstorrent/tt-metal/actions/runs/14097133289
- [x] New/Existing tests provide coverage for changes