@davorchap 2025-03-27
Great!

Will these optimizations also apply to ReduceScatter @sjameelTT ?