@sjameelTT 2025-03-27
Yeah it should be applicable.