@SeanNijjar 2025-03-27
> Great!
> 
> Will these optimizations also apply to ReduceScatter @sjameelTT ?

Yes for the most port this will carry over to all CCLs although there are some changes that are op kernel specific that would need porting over (YMMV for improvements there)