@SeanNijjar 2025-03-27
Had to run perf pipeline jobs locally. Seeing some perf deltas from when I originally recorded so I need to dig into why that's happening - was there a regression after rebase? Is there a measurement delta between the tests I ran originally and the perf pipeline scripts? Does perf swing wildly and I got consistently "lucky" runs before (yikes)? TBD ... investigating