@SeanNijjar 2025-03-27
I missed one of the optimizations from all-reduce being ported to all-gather. When applied to all-gather as well, I see another 100-400ns savings per all-gather :) I will open a separate PR